requirements.txt
streamlit
pandas
pymysql
sqlalchemy
python-dotenv
cryptography
tushare
chinese_calendar
mysql-connector-python
baostock
certifi==2026.1.4
fastapi
uvicorn[standard]
psutil
